# ParallelBlockCompressedOutputStream