… src/main/java/htsjdk/samtools/util/BlockCompressedOutputStream.java
… src/main/java/htsjdk/samtools/util/ParallelBlockCompressedOutputStream.java